requirements.txt
Flask==2.3.3
openai
elevenlabs
redis
gunicorn
python-dotenv
requests